Reordering event links
We reordered links to have the soon upcoming event on top, as well as added date information. I added June to the other link as well, just for balance, sorry if this was an overreach. When the event has past, we can move the CSL25 link down and remove the date.